jquery css.html로

+++ newper_admin/src/main/java/com/newper/repository/GoodsRepo.java
@@ -3,5 +3,11 @@ package com.newper.repository;
 import com.newper.entity.Goods;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GoodsRepo extends JpaRepository<Goods, Integer> {
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+
+public interface GoodsRepo extends JpaRepository<Goods, Long> {
+
+
+
 }
